automatically backed up by learn

+++ index.js
@@ -34,6 +34,6 @@ function makeInt(n){
   parseInt('2', 10);
 }
 function preserveDecimal(n){
-  parseInt('2.22222')
+  parseFloat('2.22222')
 }
